refactor: move `activitypub`
