+++ crates/apub-api/src/route/send_note.rs
@@ -1,8 +1,8 @@
 use apub_activitypub::core::actor::Actor;
 use apub_activitypub::model::person::Person;
 use apub_activitypub::model::{activity::CreatePersonNote, context::Context, note::Note};
+use apub_kernel::activitypub::model::{generate_activity_uri, generate_note_uri};
 use apub_kernel::prelude::*;
-use apub_kernel::repository::activity::{generate_activity_uri, generate_note_uri};
 use apub_kernel::rsa_key::model::RsaVerifyingKey;
 use apub_registry::{AppRegistry, AppRegistryExt};
 use axum::{
