+++ crates/apub-kernel/src/lib.rs
@@ -1,7 +1,6 @@
 pub mod follower;
-pub mod model;
 pub mod note;
 pub mod prelude;
-pub mod repository;
+pub mod activitypub;
 pub mod rsa_key;
 pub mod user;
