--- crates/apub-kernel/src/model.rs
@@ -1,1 +1,0 @@
-pub mod activity;
