+++ crates/apub-kernel/src/prelude.rs
@@ -1,4 +1,4 @@
-pub use crate::repository::activity::ActivityRepository;
+pub use crate::activitypub::repository::ActivityRepository;
 
 pub use crate::follower::repository::FollowerRepository;
 pub use crate::rsa_key::repository::RsaKeyRepository;
