--- crates/apub-kernel/src/repository.rs
@@ -1,1 +1,0 @@
-pub mod activity;
